+++ pluslib/Support/Providers/Security.php
@@ -1,0 +1,19 @@
+<?php
+namespace pluslib\Support\Providers;
+
+use pluslib\Security\Security as _Security;
+use pluslib\Support\ServiceProvider;
+
+class Security extends ServiceProvider
+{
+  function register()
+  {
+    //
+  }
+
+  function boot()
+  {
+    //
+    _Security::init();
+  }
+}