+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `Select::first`: apply default value
 * `Database\DB`: Custom engine (mysql, ...)
+* Implement `ServiceProvider`
 
 ## [v0.0.03](https://github.com/parsa-mostafaie/pluslib/compare/v0.0.02...v0.0.03) - 8/29/2024